\name{breed.assorted}
\alias{breed.assorted}
\title{ breed.assorted - Generate a next generation based on assorted mating }
\description{
  Generate a next generation based on assorted mating
}
\usage{
  breed.assorted(population,n.child,start,stop,...)
}

\arguments{
  \item{population}{ population structure to select parents from }
  \item{n.child}{ Number of children to breed }
  \item{start}{ Begin of region to apply selection }
  \item{stop}{ End of region to apply selection }
  \item{\dots}{ Parameters for score.kinship }
  
}
\details{
  TODO
}
\author{ 
	Danny Arends \email{Danny.Arends@gmail.com}
	Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\seealso{
  TODO
}
\examples{
  #TODO
}

\keyword{methods}
