\name{breed.random}
\alias{breed.random}
\title{ breed.random - Generate a next generation based on random mating }
\description{
  Generate a next generation based on random mating
}
\usage{
  breed.random(population,n.child)
}

\arguments{
  \item{population}{ population structure to select parents from }
  \item{n.child}{ Number of children to breed }
  
}
\details{
  TODO
}
\author{ 
	Danny Arends \email{Danny.Arends@gmail.com}
	Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\seealso{
  TODO
}
\examples{
  #TODO
}

\keyword{methods}
