\name{find.parents}
\alias{find.parents}
\title{find.parents - Check for parent, child relationship in SNP data }
\description{
	Check for parent, child relationship in SNP data 
}
\usage{
  find.parents(matrix,number=100,verbose=FALSE)
}

\arguments{
  \item{matrix}{ Matrix with individuals }
  \item{number}{ Number of individuals todo }
  \item{verbose}{ Be verbose }
  
}
\details{
	TODO
}
\author{ 
	Danny Arends \email{Danny.Arends@gmail.com}
	Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\seealso{
	\itemize{
		\item \code{\link{score.kinship}} - Calculates kinshipscores based on SNP markerdata
		\item \code{\link{find.parents}} - Finds tripplets of individuals with parents from SNP markerdata
	}
}
\examples{
	#TODO
}

\keyword{methods}
