\name{multitraitQTLs}
\alias{multitraitQTLs}

\title{QTL results of the multitrait dataset}

\description{ 
  QTL results of the multitrait dataset
} 

\usage{data(multitraitQTLs)}

\format{
  An object of class \code{mqmmulti}.  See \code{\link[qtl]{mqmscan}}
  for details. 
}  

\details{
TODO
}

\references{
TODO
}
  
\source{
TODO
}

\examples{
data(multitraitQTLs)
}

\seealso{
TODO
} 

\keyword{datasets}
