\name{qtlmatrixtomqmmulti}
\alias{qtlmatrixtomqmmulti}
\title{qtlmatrixtomqmmulti - Raw numeric QTL matrix to mqmmulti object}
\description{
	Convert a raw numeric QTL matrix to mqmmulti object
}
\usage{
	qtlmatrixtomqmmulti(cross,qtlmatrix)
}

\arguments{
  \item{cross}{ An object of class \code{cross}. See \code{\link[qtl]{read.cross}} for details, by R/qtl }
  \item{qtlmatrix}{ Numeric matrix with QTL results. }
}
\details{
  Provided upon request
}
\value{
  mqmmulti object
}
\references{ 
  None yet
}
\author{ 
	Danny Arends \email{Danny.Arends@gmail.com}
	Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\note{
  Only valid \code{cross} objects can be used
}
\seealso{
	\itemize{
    \item \code{\link{scanonetomqmmulti}} - Convert a scanone object to mqmmulti object
  }
}
\examples{
#TODO
}

\keyword{iplot}
