\name{significance.eQCL}
\alias{significance.eQCL}
\title{ significance.eQCL - Print out thresholds for an eQTL permutation object }
\description{
  Summarizes the results and prints out thresholds for an eQTL permutation object
}
\usage{
  significance.eQCL(eQCLperm, alpha, verbose = FALSE)
}

\arguments{
  \item{eQCLperm}{ Object from the \code{\link[iqtl]{diffCorPermutation}} }
  \item{alpha}{ Alpha level for hotspot detection }
  \item{verbose}{ Be verbose }
}
\details{
  TODO
}
\author{ 
	Danny Arends \email{Danny.Arends@gmail.com}
	Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\seealso{
  TODO
}
\examples{
  #TODO
}

\keyword{methods}
