\name{significanceThresholds}
\alias{significanceThresholds}
\title{ significanceThresholds - Print out thresholds for differential correlation }
\description{
  Summarizes the results and prints out thresholds for differential correlation
}
\usage{
  significanceThresholds(difCntPerm)
}

\arguments{
  \item{difCntPerm}{ Object from the \code{\link[iqtl]{diffCorPermutation}} }
}
\details{
  TODO
}
\author{ 
	Danny Arends \email{Danny.Arends@gmail.com}
	Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\seealso{
  TODO
}
\examples{
  #TODO
}

\keyword{methods}
