\name{iqtl-internal}

%zzz.R
\alias{has_d}
\alias{has_qtlHD}
\alias{get_IqtlHDEnv}

%iqtlhd.utils.R
\alias{connected}

\alias{contrastqtlmapping.internal}
\alias{donothing}
\alias{reciproce}
\alias{inverseF}
\alias{findperiod}
\alias{scaledowncontrast}
\alias{scaledowncontrasts}
\alias{XGAPheader}
\alias{XGAPread}
\alias{DownloadnSave}
\alias{Generate_Statement}
\alias{Names}
\alias{Rn}
\alias{RnM}
\alias{RnMN}
\alias{RnN}
\alias{RnV}
\alias{RnVN}
\alias{SNPMarker}
\alias{analyseQTL}
\alias{analysis.differentialCorrelation}
\alias{assortedmating}
\alias{assortedmating.init}
\alias{avgResults}
\alias{checkcutoffndrop}
\alias{client.start}
\alias{countVectorThreshold}
\alias{countDifCorThreshold}
\alias{correlationmatrix}
\alias{create.client}
\alias{create.cluster}
\alias{createdatafile}
\alias{creategnuplot}
\alias{difCorTargets}
\alias{imageDifCnt}
\alias{summaryDifCnt}
\alias{createheatmapbymarkerscript}
\alias{createprofilebymarkerscript}
\alias{datacheck}
\alias{create.child}
\alias{est_runtime}
\alias{examples}
\alias{generateESTfile}
\alias{generatePERMfile}
\alias{generateQTLfile}
\alias{generateRunfile}
\alias{generateSIF}
\alias{getraw}
\alias{inference}
\alias{install_libraries}
\alias{loadAvgResults}
\alias{lodtopval}
\alias{make.child}
\alias{mapQTL}
\alias{mapQTLs}
\alias{modelQTL}
\alias{mqmmultitomatrix}
\alias{mqmsupportint}
\alias{myNames}
\alias{ourcat}
\alias{permQTL}
\alias{plot.all}
\alias{plot.combined}
\alias{plot.histkinship}
\alias{prepare_cluster}
\alias{pvaltolod}
\alias{report}
\alias{smoothV}
\alias{test.cluster}
\alias{test.differentialCorrelation}
\alias{testdata}
\alias{testgnuplot}
\alias{traitspeaked}


\alias{StringArrayToD}
\alias{compareTraits}
\alias{compareTraits.test}
\alias{correctQTLeffect}
\alias{correcttrait}
\alias{createRrepos}
\alias{endlessloop}
\alias{gcLoop}
\alias{geneMergeGO}
\alias{getLine}
\alias{getMarkerAnnotation}
\alias{getNCBIfiles}
\alias{hasSpecies}
\alias{highcpu}
\alias{individualcorrection}
\alias{loop}
\alias{markereffect}
\alias{markereffects}
\alias{mem}
\alias{nocpu}
\alias{numLines}
\alias{parseFTPdir}
\alias{read.fasta}
\alias{read.genome}
\alias{searchNCBIgenome}
\alias{test.NCBIgenome}
\alias{test.binary}
\alias{threads}
\alias{write.cross.binary}

\title{Internal iqtl functions}

\description{
  Internal iqtl functions. These are generally not to be called by the
  user.
}

\author{Danny Arends, \email{Danny.Arends@Gmail.com} }

\keyword{internal}
